backend/requirements.txt
fastapi
uvicorn
tensorflow
torch
torchvision
numpy
pillow
openai
dotenv